feat(admin): 관리자 남은 api 개발

+++ src/main/java/com/example/ormi5finalteam1/service/AdminService.java
@@ -2,13 +2,20 @@ package com.example.ormi5finalteam1.service;
 
 import com.example.ormi5finalteam1.common.exception.BusinessException;
 import com.example.ormi5finalteam1.common.exception.ErrorCode;
-import com.example.ormi5finalteam1.domain.comment.Comment;
+import com.example.ormi5finalteam1.domain.comment.dto.AdminCommentDto;
 import com.example.ormi5finalteam1.domain.post.Post;
+import com.example.ormi5finalteam1.domain.post.dto.AdminPostDetailDto;
+import com.example.ormi5finalteam1.domain.post.dto.AdminPostListDto;
+import com.example.ormi5finalteam1.domain.user.Role;
 import com.example.ormi5finalteam1.domain.user.User;
+import com.example.ormi5finalteam1.domain.user.dto.UserInfoDto;
 import com.example.ormi5finalteam1.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @Transactional
@@ -19,10 +26,16 @@ public class AdminService {
     private final PostService postService;
     private final CommentService commentService;
 
+    public List<UserInfoDto> getAllUserList() {
+
+        return userRepository.findAllByOrderByRoleAscId().stream()
+                .map(UserInfoDto::toDto).collect(Collectors.toList());
+    }
     public void changeUserStatus(Long userId) {
 
         User normalUser = userRepository.findById(userId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.USER_NOT_FOUND));
+        if (normalUser.getRole() == Role.ADMIN) throw new BusinessException(ErrorCode.HAS_ADMIN_AUTHORITY);
         if(normalUser.getDeletedAt() == null) {
             if (normalUser.isActive()) normalUser.deactivateUser();
             else normalUser.activateUser();
@@ -33,6 +46,7 @@ public class AdminService {
 
         User normalUser = userRepository.findById(userId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.USER_NOT_FOUND));
+        if (normalUser.getRole() == Role.ADMIN) throw new BusinessException(ErrorCode.HAS_ADMIN_AUTHORITY);
         if (normalUser.getDeletedAt() == null) normalUser.delete();
         else throw new BusinessException(ErrorCode.ALREADY_DELETED);
     }
@@ -45,5 +59,16 @@ public class AdminService {
 
     public void deleteComment(Long postId, Long commentId) {
         commentService.deleteCommentByAdmin(postId, commentId);
+    }
+
+    public List<AdminPostListDto> getAllPostList() {
+
+        return postService.getAllPostsByAdmin().stream().map(AdminPostListDto::toDto).collect(Collectors.toList());
+    }
+
+    public AdminPostDetailDto getPostById(Long postId) {
+
+        return AdminPostDetailDto.toDto(postService.getPostByAdmin(postId),
+                commentService.getCommentsByPostId(postId).stream().map(AdminCommentDto::toDto).toList());
     }
 }
